requirements.txt
django
graphene-django
pymongo